project: verify the largest files @complete (settings, audio, main VC, store, download)
Disassembly-verified sweep (agent-assisted, spot-checked) of the biggest files.
UserSettingData (121), AudioManager (69), DownloadMain (65), StoreMainViewController
(64) and MainViewController (80), plus resolution of the remaining partials.

Real bugs found and fixed:
- UserSettingData: 18 save-store fixes — wrong NSUserDefaults key strings
  (IsFriendSelected, IsPopnLinkSelected, the misspelled IsPolicyAccesped,
  ConvertCode, IsBemaniCollaboOpen), missing value clamps (seVolume/touchSoundVolume/
  musicSort lower bound, treasurePoint cap 9999, invite<0), and bias/bit fixes
  (openedLoginBonusId +/-1, haveTouchSoundFlg |1, touchSoundKind int16 clamp,
  saveDate: read-compare-synchronise). These affect save compatibility.
- AudioManager: SeInstance struct layout — group is at +0x8 (12-byte stride), not
  +0x4, silently correcting every SE method that indexes the struct. resumePlayer
  fades BGM back over 0.3s; setSeVolume group-0 loops 8 per-voice gain sets and
  forwards the raw int level (the /127 is inside setGroupVolume); playSe:Volume: is
  self-contained (clamp int level, no m_seVolume write). Retyped the group-volume
  signatures to int.
- DownloadMain: ContextType "application/json"; server-clock date separator "-";
  news-link separator "@NEWSLINK=".
- MapSelectSplitVC init clamp/arrow displacement; AcNoteMng setupLaneMapping exact
  100000-pass shuffle; AcMainTask sugoroku fixes (kill-flag gate, fastRecord +0x4c,
  squareFrameIdx float). Reworded interface-declaration address false-positives.

MainViewController render/task frame loop and neCreateOrthoViewport/neSetCurrentViewport
frame-begin wiring verified faithful. Overall coverage now 94%.

+++ Project/System/src/Sound/AudioManager.h
@@ -114,7 +114,7 @@ static const RSND_INSTANCE_ID RSND_INSTANCE_ID_ERROR = (RSND_INSTANCE_ID)-1;
                 resourceId:(RSND_SOURCE_ID)resourceId; // Ghidra: @ 0x1f234
 - (RSND_INSTANCE_ID)playSe:(NSString *)name
                 resourceId:(RSND_SOURCE_ID)resourceId
-                    Volume:(float)volume;           // Ghidra: @ 0x1f2d8
+                    Volume:(int)volume;             // Ghidra: @ 0x1f2d8
 - (BOOL)stopSe:(RSND_INSTANCE_ID)instanceId;        // Ghidra: @ 0x1f3d0
 - (BOOL)stopSeAll;                                  // Ghidra: @ 0x1f630
 - (void)setSeVolume:(int)volume groupId:(int)group; // Ghidra: @ 0x1f99c
